Moved OrderedCode to be package protected in org/apache/beam/sdk/io/gcp/spanner package.
